requirements.txt
matplotlib
numpy
pillow
wordcloud
